chore: Renaming and seperating components in calandar grid to allow for another yearly-grid that shares functionality
