+++ src/date-range-picker/calendar/grids/index.tsx
@@ -11,7 +11,7 @@ import { hasValue } from '../../../internal/utils/has-value';
 import InternalSpaceBetween from '../../../space-between/internal';
 import { DateRangePickerProps, DayIndex } from '../../interfaces';
 import { findDateToFocus } from '../utils';
-import { Grid } from './grid';
+import { MonthlyGrid } from './monthly-grid';
 
 import styles from '../../styles.css.js';
 
@@ -168,7 +168,7 @@ export const Grids = ({
     <div ref={containerRef} onFocus={onGridFocus} onBlur={onGridBlur}>
       <InternalSpaceBetween size="xs" direction="horizontal">
         {!isSingleGrid && (
-          <Grid
+          <MonthlyGrid
             className={styles['first-grid']}
             baseDate={addMonths(baseDate, -1)}
             selectedEndDate={selectedEndDate}
@@ -188,7 +188,7 @@ export const Grids = ({
             ariaLabelledby={`${headingIdPrefix}-prevmonth`}
           />
         )}
-        <Grid
+        <MonthlyGrid
           className={styles['second-grid']}
           baseDate={baseDate}
           selectedEndDate={selectedEndDate}
